selectively disable the libpulsedsp DSO and padsp wrapper
The libpulsedsp DSO is not compatible with 64-bit time_t, and is a wrapper
for compatibility with an API that's basically obsolete for 15 years.
And nowadays pulseaudio itself is largely superseded by pipewire.

padsp is mainly useful for legacy i386 binaries and maybe
a few very early legacy amd64 binaries (late 90s/early 00s era games,
especially ported by Loki Games) which might assume OSS audio, and not
support ALSA or PulseAudio directly. Since we're breaking the platform ABI
on armel/armhf/etc. *anyway*, we can't rely on the ability to run legacy
binaries even if they exist, so it's probably no longer interesting to
provide padsp on those architectures.

However, Debian's i386 is sticking with the old 32-bit time_t.

Forwarded: not-needed