Add patch to support old webrtc-audio-processing library
We have webrtc-audio-processing v0.3 in Debian unstable. There's a v1.0
packaged in experimental, but it failed to build on i386, mips64el and
riscv64 so is not fit for release. The binary package names for that
will need to be changed for the "-1" version suffix in new versions, so
it could have to go through the NEW queue.

Instead of waiting for that, add a patch that tries to add build-time
compatibility with both versions so that we can have a smoother
transition to the new library when it's available in unstable.