Add patch to fix segfault with too many UCM devices
Reportedly, PulseAudio crashes with a segfault when trying to use MOTU
USB sound cards. The root of the issue is that a piece of recursive code
blows up with 2^n calls (n being around 20 devices) where it was assumed
to be tail-call optimizable. Add a patch that converts it to iterative
form, which is also recently applied in Arch Linux.